+++ hoohoo_web/src/Pages/Partnership/B2B/sections/PartnershipShop.tsx
@@ -4,7 +4,7 @@ import { useLanguage } from '../../../../components/hooks/LanguageContext';
 import Wrapper from '../../../../components/Wrapper/Wrapper';
 import { emailTo } from '../../../../util/email';
 import { logButtonEvent, PageName } from '../../../../util/firebase_custom_event';
-import { HomeTransitionButton } from '../../../Home/HomeIntroPage';
+import { HomeTransitionButton } from '../../../Home/styles';
 import { ButtonBox } from '../PhotoVideoes';
 import {
   PartnershipContainer,
