+++ hoohoo_web/src/Pages/Partnership/ticketeer/AskTicketeer.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
-import {HomeTransitionButton} from '../../Home/HomeIntroPage';
-import {ButtonBox} from '../B2B/PhotoVideoes';
+
+import { HomeTransitionButton } from '../../Home/styles';
+import { ButtonBox } from '../B2B/PhotoVideoes';
 const Container = styled.section`
   width: 100%;
   background-color: transparent;
