+++ hoohoo_web/src/lang/ko.json
@@ -4,7 +4,7 @@
     "data2": {}
   },
   "Nav": {
-    "logo": {"text": "어스메라", "image": "/Images/icon_image.png", "link": "/"},
+    "logo": {"text": "어스메라", "image": "/Images/icon_image.png", "link": "/ko/"},
     "lang": ["한", "영"],
 
     "navlist": [
